requirements.txt
pyyaml==3.12
mock==2.0.0
